requirements.txt
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
idna==2.6
Mastodon.py==1.2.2
pkg-resources==0.0.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
